Extract LanguageModal drag-to-reorder into a composable
Addresses review feedback that the drag logic cluttered the modal. Moved
into useDragReorder.ts, and simplified along the way: since startDrag
already calls setPointerCapture, pointermove/pointerup/pointercancel can
be plain v-on directives on the handle instead of manually added/removed
window listeners.
